requirements.txt
bleach==5.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
docutils==0.19
glfw==2.5.4
idna==3.3
importlib-metadata==4.12.0
keyboard==0.13.5
keyring==23.8.2
Pillow==9.2.0
pkginfo==1.8.3
Pxucz==0.0.1.2
Pygments==2.12.0
PyOpenGL==3.1.6
PyOpenGL-accelerate==3.1.6
pywin32-ctypes==0.2.0
readme-renderer==36.0
requests==2.28.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rich==12.5.1
six==1.16.0
twine==4.0.1
urllib3==1.26.11
webencodings==0.5.1
zipp==3.8.1
